+++ CHANGES.md
@@ -6,6 +6,7 @@ next
 
 * Dropped ServiceApi struct in favor of using Session directly.
 * Rework of the whole compute module.
+* IdentityAuthMethod renamed to PasswordAuth.
 
 0.0.8
 -----